Rework validation
All validation is now "context" aware. Essentially, Result -> Context[T]  and
the Context exposes an Env T. This is so validation can be done using context-
specific data. For example, on signup, we might require a Name field based on
some Tenant configuration.

The "Context" is also more stateful than the previous Result. It knows the
current field being validated and thus, a lot of the API has been streamlined.
Specifically, the *Field is largely an internal detail, and it's no longer
exposed/needed in custom validation functions.

Similarly, the Context also has a reference to the full Input and the current
Object, so these no longer have to be passed aroud.

In total, we've gone from passing 4 fields around (Result, Input, Object and
Field) to just needing the Context.

Also reworked most validators to inline the logic. For example, the
StringValidator no long has an array of validators. Instead, it's directly
configured with all the validation settings (min/max length, pattern, choices,
etc) and validates directly. This is faster and will make it easier to serialize
/deserialize validators.